+++ BettingTest/Metrics/HomeAdvantageMetric.t.cs
@@ -165,7 +165,7 @@ namespace BettingTest
             HomeAdvantageMetric metric = new HomeAdvantageMetric(metricConfig, year, configManagerMock.Object, fixtureRetrieverMock.Object);
 
             // Act
-            metric.GetPoints(out int pTeam1, out int pTeam2, teamId1, teamId2, actualFixture);
+            metric.GetPoints(out double pTeam1, out double pTeam2, teamId1, teamId2, actualFixture);
 
             // Assert
             Assert.AreEqual(pTeam1, 1);
@@ -184,7 +184,7 @@ namespace BettingTest
             HomeAdvantageMetric metric = new HomeAdvantageMetric(metricConfig, year, configManagerMock.Object, fixtureRetrieverMock.Object);
 
             // Act
-            metric.GetPoints(out int pTeam1, out int pTeam2, teamId1, teamId2, actualFixture);
+            metric.GetPoints(out double pTeam1, out double pTeam2, teamId1, teamId2, actualFixture);
 
             // Assert
             Assert.AreEqual(pTeam1, 1);
